{{cookiecutter.package_name}}/requirements/test.txt
-r base.txt

pytest
pytest-cov
pytest-django
pytest-sugar
pytest-xdist
